# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Score API",
        "contact": {
            "name": "Score API Developer Team",
            "url": "https://github.com/pine09/ScoreAPI"
        },
        "version": "1.0"
    },
    "host": "score.app",
    "basePath": "",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/v1/admin/konsentrasi": {
            "get": {
                "tags": [
                    "Admin on Konsentrasi"
                ],
                "summary": "Mengambil semua Konsentrasi yang ada.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data Konsentrasi. (max 5 per page)",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/konsentrasi"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized action."
                    },
                    "403": {
                        "description": "Forbidden action."
                    }
                }
            },
            "post": {
                "tags": [
                    "Admin on Konsentrasi"
                ],
                "summary": "Mendaftarkan Konsentrasi baru.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Data Konsentrasi Baru.",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/konsentrasi"
                            }
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data Konsentrasi baru."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    },
                    "403": {
                        "description": "Forbidden action."
                    },
                    "409": {
                        "description": "Conflict."
                    },
                    "422": {
                        "description": "Unprocessable Entity."
                    }
                }
            }
        },
        "/api/v1/admin/konsentrasi/{id}": {
            "get": {
                "tags": [
                    "Admin on Konsentrasi"
                ],
                "summary": "Mengambil data Konsentrasi spesifik",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data Konsentrasi yang dicari"
                    },
                    "401": {
                        "description": "Unauthorized action."
                    },
                    "403": {
                        "description": "Forbidden action."
                    },
                    "404": {
                        "description": "Resource not found."
                    }
                }
            },
            "put": {
                "tags": [
                    "Admin on Konsentrasi"
                ],
                "summary": "Mengedit data Konsentrasi.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Data Konsentrasi (Edit).",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/konsentrasi"
                            }
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data Konsentrasi edited."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    },
                    "403": {
                        "description": "Forbidden action."
                    },
                    "404": {
                        "description": "Resource not found."
                    },
                    "409": {
                        "description": "Conflict."
                    },
                    "422": {
                        "description": "Unprocessable Entity."
                    },
                    "500": {
                        "description": "Internal Server Error."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Admin on Konsentrasi"
                ],
                "summary": "Men-delete data Konsentrasi.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data Konsentrasi deleted."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    },
                    "403": {
                        "description": "Forbidden action."
                    },
                    "404": {
                        "description": "Resource not found."
                    },
                    "422": {
                        "description": "Unprocessable Entity."
                    },
                    "500": {
                        "description": "Internal Server Error."
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "konsentrasi": {
            "properties": {
                "konsentrasi_code": {
                    "type": "string"
                },
                "konsentrasi_name": {
                    "type": "string"
                },
                "jurusan_id": {
                    "type": "integer",
                    "format": "int32"
                }
            }
        },
...
    }
}